adapt upgrade to wflow v1 sediment
